Horse XP, Taming
Added:
Horse taming + config for how hard it is to tame them, if only trainers can tame, train and sell, where to sell, how much to sell for
Horse XP: i fixed a booboo i made where it wouldnt check if you were still on the horse before giving xp.
Horse labels: When you get on a horse, it will tell you what it is. Config toggle for that too.